requirements.txt
 anthropic>=0.25.0
  pyyaml>=6.0
  tqdm>=4.66
  rich>=13.7
  python-dotenv>=1.0
  pytest>=8.0